+++ R/TROLLv3_sim.R
@@ -1,0 +1,10 @@
+#' TROLL sim
+#'
+#' A TROLL outputs from a 1-year simulation on a 100x100 grid with other parameters 
+#' are sets to default parameters and using TROLLv3_species, TROLLv3_climatedaytime12, and
+#' TROLLv3_daytimevar for use in tests and examples. Ecosystem level output has
+#' been thinned and species output has been removed to save disk space.
+#'
+#' @format A trollsim object.
+#'
+"TROLLv3_sim"