+++ R/generateParam.R
@@ -61,7 +61,7 @@ NULL
 #' trollDAETest <- setupDesign(paramsBounds = paramsBounds,
 #' ntotalsim = 10,echo = FALSE)
 #' 
-#' generate_params(obj = trollDAETest,nyearsInit = 600)
+#' generate_params(obj = trollDAETest,nyearsInit = 600, fnSpecies = fnSpecies)
 #' 
 #' @name generate_params
 NULL