+++ R/processSim.R
@@ -1,7 +1,9 @@
+#' @include TROLLv3_sim.R
 #' @importFrom parallel detectCores
 #' @importFrom tidyr unnest
 #' @importFrom dplyr filter_at matches vars all_vars any_vars
-#' @importFrom hetGP IMSPE_optim
+#' @import hetGP
+#' @importFrom stats var cor cov quantile qt
 NULL
 
 #' A function to process simulation according to 
@@ -16,19 +18,21 @@ NULL
 #' convergence at the tail of simulation.
 #' @param cores int. Number of cores to use.
 #' @param filesave chr. Path to save dae in processing.
+#' @param targets list. 
 #' @param nyearsStep int. Number of year to add 
 #' if convergence is not reached.
 #' @param nyearsMax int. Maximum number of year 
 #' to reach convergence.
 #' @param simOpts list for options about the surrogate modeling procedure.
-#' (see \code{hetGP::\link[hetGP:mleHetGP]{mleHetGP}} for details)
+#' (see \code{\link[trollcalibr:computeGP]{computeGP}} and 
+#' \code{\link[trollcalibr:computeHM]{computeHM}} for details)
 #' @param verbose cat processing infos.
 #'
 #' @return a trolldae object 
 #' @export
 #'
 #' @examples
-#' 
+#' \dontrun{
 #'  set.seed(123)
 #'  require(tibble)
 #'  require(dplyr)
@@ -42,7 +46,7 @@ NULL
 #'    type = c("global","global", "experiment"))
 #'    
 #' DAE <- setupDesign(name = "test",paramsBounds = paramsBounds,ntotalsim = 200,  
-#' nreplica = 4, sequential = TRUE,ninitsim = 25)
+#' nreplica = 4, sequential = TRUE,ninitsim = 50)
 #'    
 #' DAEwithParams <- generate_params(DAE, nyearsInit = 2)
 #' 
@@ -80,17 +84,19 @@ NULL
 #'  
 #'  processSim(DAEwithExp,
 #'  checkConverge = FALSE,
-#'  inputs = list("exInputs" = NULL),
+#'  inputs = list("exInputs" = NULL), 
 #'  nyearsSampling = 1,
 #'  cores = 2,
 #'  filesave = getwd())
+#'  }
 #'
 processSim <- function(dae,
                        inputs,
                        checkConverge = TRUE,
                        nyearsSampling,
                        cores,
-                       filesave,
+                       filesave = NULL,
+                       targets = NULL,
                        nyearsStep = 0L,
                        nyearsMax = NULL,
                        simOpts = NULL,
@@ -101,10 +107,12 @@ processSim <- function(dae,
     stop("'dae' argument of 'processSim' must be a trolldae")
   }
   
-  if (!is(filesave, c("character")) || !dir.exists(filesave)) {
-    stop("'filesave' argument of 'processSim' must be a character and dir in path must exist")
+  if (!is.null(filesave)) {
+    if (!is(filesave, c("character")) || !dir.exists(filesave)) {
+      stop("'filesave' argument of 'processSim' must be a character and dir in path must exist")
+    }
   }
-  
+
   if (!is(checkConverge, c("logical"))) {
     stop("'checkConverge' argument of 'processSim' must be a logical")
   }
@@ -145,9 +153,26 @@ processSim <- function(dae,
     nyearsMax <- dae@simopts$nyearsInit+1
   }
   
+  if(!is.null(targets)){
+    if (!is(targets,"list")) {
+      stop("'targets' argument of 'processSim' must be a list")
+    }
+    
+    if (is.null(dae@experiments[[1]]@outputs.opts$summary)) {
+      stop("'targets' argument of 'processSim' cannot be set for experimental setup without summary")
+    }
+    
+    if(!all(names(targets) %in% colnames(dae@experiments[[1]]@outputs.opts$summary))){
+      stop("'targets' argument of 'processSim' must include all outputs ranges")
+    }
+  } 
   
-  if (is.null(simOpts) && dae@type == "GP") {
-    simOpts <- list("lower" = NULL,
+  if (is.null(simOpts) && dae@type[1] != "RAW") {
+    simOpts <- list("genPts" = dae@doeopts$nsim,
+                    "ratioVal" = 0.2,
+                    "nbWave" = 2,
+                    "targets" = targets,
+                    "lower" = NULL,
                     "upper" = NULL,
                     "covtype" = NULL,
                     "noiseControl" = NULL,
@@ -157,9 +182,11 @@ processSim <- function(dae,
   IDsim <- blocki <- ID_iter <- iter <- NULL
   globalData <- speciesData <- climateData <- NULL
   dailyData <- lidarData <- forestData <- NULL
-  sim_stack <- NULL
+  sim_stack <- . <- nsim <- simulation <- NULL
   
   path <- dae@path
+  
+  dae@state <- "In-processing"
   
   setup <- dae@experiments[[1]]
   
@@ -175,6 +202,8 @@ processSim <- function(dae,
   
   nrep <- params %>% select(IDsim) %>% 
     mutate(ID_iter = ceiling(row_number()/(cores)))
+  
+  dae@doeopts <- c(dae@doeopts, list("checkConverge" = if(checkConverge){data.frame("IDsim" = nrep$IDsim, nsim = NA)}else{NULL}))
   
   dae@simopts$settingsGP <- simOpts
   
@@ -258,11 +287,11 @@ processSim <- function(dae,
           sim_stack_tmp <- sim_stack
         }
         
-        
-        
-        
+
         
         if (checkConverge) {
+          
+          dae@doeopts$checkConverge$nsim[which(dae@doeopts$checkConverge$IDsim == simulations)] <- nsim
           
           expSummary <- createExp(id = 1,type = "Summary",
                                   deltaT = 1L,
@@ -323,7 +352,56 @@ processSim <- function(dae,
         
       }
       rm(sim_stack,sim_stack_tmp,listSimi,allconverged)
+      
+      if (blocki %% 10 == 0 && !is.null(filesave)) {
+        save(dae,inputs, 
+             checkConverge,
+             nyearsSampling,
+             cores,
+             filesave,
+             targets,
+             nyearsStep,
+             nyearsMax,
+             simOpts, 
+             file = paste0(filesave,
+                           "/save_iter", blocki , "-", max(nrep$ID_iter), "_", 
+                           gsub(":", "-",
+                                gsub(
+                                  " ", "_",
+                                  timestamp(
+                                    prefix = "",
+                                    suffix = "",
+                                    quiet = T
+                                  )
+                                ))
+                           ,".rda"))
+      }
     }
+    
+    if (!is.null(filesave)) {
+      save(dae,inputs, 
+           checkConverge,
+           nyearsSampling,
+           cores,
+           filesave,
+           targets,
+           nyearsStep,
+           nyearsMax,
+           simOpts, 
+           file = paste0(filesave,
+                         "/save_DAE_", 
+                         gsub(":", "-",
+                              gsub(
+                                " ", "_",
+                                timestamp(
+                                  prefix = "",
+                                  suffix = "",
+                                  quiet = T
+                                )
+                              ))
+                         ,".rda"))
+    }
+    
     
     summarySim <- NULL
     outputsSim <- NULL
@@ -357,7 +435,15 @@ processSim <- function(dae,
     
     rm(summarySim,outputsSim,listSim)
   }
-  switch (dae@type,
+  switch (dae@type[1],
+          "HM" = {dae <- .historyMatching(dae = dae,
+                                          simOpts = simOpts,
+                                          cores = cores,
+                                          checkConverge = checkConverge,
+                                          filesave = filesave,
+                                          nyearsSampling = nyearsSampling,
+                                          nyearsStep = nyearsStep,
+                                          nyearsMax = nyearsMax)},
           "GP" = {dae <- .seqDesignGP(dae = dae,
                                       simOpts = simOpts,
                                       cores = cores,
@@ -369,13 +455,15 @@ processSim <- function(dae,
           "RAW" = {})
   
   dae@state <- "Post-simulation"
-  
+
   return(dae)
 }
 
 .fnSummary <- function(x,parameters,inputs,...){
   iter <- sum10 <- sum30 <-  NULL
   dens10 <- dens30 <- agb <- NULL
+  
+  data(TROLLv3_sim,envir = environment())
   
   if (identical(x,TROLLv3_sim)) {
     inputs = list("nyearsSim" = 100,
@@ -465,7 +553,7 @@ processSim <- function(dae,
   
   nseqsim <- dae@doeopts$ntotalsim - dae@doeopts$nsim
   
-  horizons <- lapply(seq_len(dim(dae@ysim)[2]),FUN =  function(x){rep(0, )}) 
+  horizons <- lapply(seq_len(dim(dae@ysim)[2]),FUN =  function(x){rep(0, nseqsim)}) 
   
   GP <- computeGP(dae = dae,
                   lower = simOpts$lower,
@@ -558,6 +646,8 @@ processSim <- function(dae,
     
     daeWIP <- addExp(daeWIP,setup)
     
+    daeWIP@state <- "In-processing"
+    
     daeWIP <- processSim(dae = daeWIP,
                          inputs = dae@experiments[[1]]@inputs.opts,
                          checkConverge = checkConverge,cores = cores,
@@ -575,10 +665,10 @@ processSim <- function(dae,
     dae@ysim <- rbind(dae@ysim,daeWIP@ysim)
     
     GP@models <-  lapply(seq_len(dim(dae@ysim)[2]),
-                         function(x){update(GP@models[[x]], 
+                         function(x){eval(parse(text = "update(GP@models[[x]], 
                                             Xnew=daeWIP@lhs, 
                                             Znew=daeWIP@ysim, 
-                                            ginit=GP@models[[x]]$g*1.01)})
+                                            ginit=GP@models[[x]]$g*1.01)"))})
     
     if(seqIter %% 25 == 0){
       GP <- computeGP(dae = dae,
@@ -590,6 +680,101 @@ processSim <- function(dae,
     }
     
   }
+  
+  dae@simopts$settingsGP <- 
+  
+  return(dae)
+}
+
+.historyMatching <- function(dae,
+                             simOpts,
+                             cores,
+                             checkConverge,
+                             nyearsSampling,
+                             filesave,
+                             nyearsStep,
+                             nyearsMax){
+  
+  HMWIP <- computeHM(dae = dae,
+                     targets = simOpts$targets,
+                     genPts = floor(dae@doeopts$nsim/(1-simOpts$ratioVal)),
+                     ratioVal = simOpts$ratioVal,
+                     inmem = FALSE)
+  
+  
+  
+  for (wave in seq_len(simOpts$nbWav)) {
+    cat(paste0("Computing TROLL history matching simulation # ", wave , " / ",
+               simOpts$nbWav, "\n",
+               gsub(":", "-",
+                    gsub(
+                      " ", "_",
+                      date()
+                    )),"\n"))
+    
+    daeWIP <- trolldae(name = dae@name,
+                       boundaries = dae@boundaries,
+                       doeopts = dae@doeopts,
+                       state = dae@state,path = dae@path,
+                       simopts = simOpts,type = "RAW")
+    
+    
+    
+    X <- rbind(HMWIP@modelsopts$newLHS, 
+                 matrix(rep(t(HMWIP@modelsopts$newLHS), max(daeWIP@doeopts$nreplica -1, 0) ), 
+                        ncol= dim(HMWIP@modelsopts$newLHS)[2] , byrow=TRUE))
+    daeWIP@lhs <- X[sample(nrow(X)),]
+    
+    
+    daeWIP@doeopts$nsim <- dim(daeWIP@lhs)[1]/daeWIP@doeopts$nreplica
+    
+    daeWIP <- generate_params(obj = daeWIP,
+                              nyearsInit = dae@simopts$nyearsInit,
+                              fnGlobal = dae@doeopts$fnParams$fnGlobal,
+                              fnClimate = dae@doeopts$fnParams$fnClimate,
+                              climateInit = dae@doeopts$fnParams$climateInit,
+                              fnDaily = dae@doeopts$fnParams$fnDaily,
+                              dailyInit = dae@doeopts$fnParams$dailyInit,
+                              fnSpecies = dae@doeopts$fnParams$fnSpecies,
+                              speciesInit = dae@doeopts$fnParams$speciesInit,
+                              fnLidar = dae@doeopts$fnParams$fnLidar,
+                              covariates = dae@doeopts$fnParams$covariates,
+                              echo = FALSE, minID = dim(dae@ysim)[1])
+    
+    setup <- setupExperiments(dae = daeWIP,
+                              listexp = dae@experiments[[1]]@listexp[2:length(dae@experiments[[1]]@listexp)],
+                              inputs = dae@experiments[[1]]@inputs.opts)
+    
+    daeWIP <- addExp(daeWIP,setup)
+    
+    daeWIP@state <- "In-processing"
+    
+    daeWIP <- processSim(dae = daeWIP,
+                         inputs = dae@experiments[[1]]@inputs.opts,
+                         checkConverge = checkConverge,cores = cores,
+                         nyearsSampling = nyearsSampling,filesave = filesave,
+                         nyearsStep = nyearsStep, nyearsMax = nyearsMax, simOpts = simOpts,
+                         verbose = FALSE)
+    
+    dae@lhs <- rbind(dae@lhs, daeWIP@lhs)
+    
+    dae@params <- rbind(dae@params, daeWIP@params)
+    
+    dae@doeopts$nsim <- dae@doeopts$nsim + daeWIP@doeopts$nsim
+    
+    dae@xsim <- rbind(dae@xsim,daeWIP@xsim)
+    dae@ysim <- rbind(dae@ysim,daeWIP@ysim)
+    
+    HMWIP <- computeHM(dae = daeWIP,
+                       targets = simOpts$targets,
+                       genPts = floor(dae@doeopts$nsim/(1-simOpts$ratioVal)),
+                       ratioVal = simOpts$ratioVal,
+                       inmem = FALSE)
+    
+  }
+  
+  
+  dae@type <- c("GP")
   
   return(dae)
 }