+++ R/surcalib.R
@@ -6,8 +6,10 @@ NULL
 #' @slot yobs d.f. Observed summary statitics.
 #' @slot xobs d.f. Observed variables for available summary statitics.
 #' @slot xcalib d.f. Available variable on the whole studied setup.
-#' @slot params.formulas list. list of equations linking estimated parameters.
-#' @slot cov.formulas d.f. Available covariable for 'params.formulas' 
+#' @slot ycalib d.f. Estimated summary statitics on the whole studied setup.
+#' @slot paramCalib d.f Estimated parameters values on the whole studied setup.
+#' @slot paramsFormulas list. list of equations linking estimated parameters.
+#' @slot covFormulas d.f. Available covariable for 'paramCalib' 
 #' on the whole studied setup.
 #' @slot surmodel list. an adjusted TROLL surrogate model.
 #' @slot calib.opts list. Calibrations options.
@@ -18,15 +20,19 @@ setClass(
   representation(yobs = "data.frame",
                  xobs = "data.frame",
                  xcalib = "data.frame",
-                 params.formulas = "list",
-                 cov.formulas = "data.frame",
+                 ycalib = "data.frame",
+                 paramCalib = "data.frame",
+                 paramsFormulas = "list",
+                 covFormulas = "data.frame",
                  surmodel = "list",
                  calib.opts = "list"),
   prototype(yobs = data.frame(),
             xobs = data.frame(),
             xcalib = data.frame(),
-            params.formulas = list(),
-            cov.formulas = data.frame(),
+            ycalib = data.frame(),
+            paramCalib = data.frame(),
+            paramsFormulas = list(),
+            covFormulas = data.frame(),
             surmodel = list(),
             calib.opts = list())
 )
@@ -36,8 +42,10 @@ setClass(
 #' @param yobs d.f. Observed summary statitics.
 #' @param xobs d.f. Observed variables for available summary statitics.
 #' @param xcalib d.f. Available variable on the whole studied setup.
-#' @param params.formulas list. list of equations linking estimated parameters.
-#' @param cov.formulas d.f. Available covariable for 'params.formulas' 
+#' @param ycalib d.f. Estimated summary statitics on the whole studied setup.
+#' @param paramCalib d.f Estimated parameters values on the whole studied setup.
+#' @param paramsFormulas list. list of equations linking estimated parameters.
+#' @param covFormulas d.f. Available covariable for 'paramCalib' 
 #' on the whole studied setup.
 #' @param surmodel list. an adjusted TROLL surrogate model.
 #' @param calib.opts list. Calibrations options.
@@ -47,16 +55,20 @@ setClass(
 surcalib <- function(yobs = data.frame(),
                      xobs = data.frame(),
                      xcalib = data.frame(),
-                     params.formulas = list(),
-                     cov.formulas = data.frame(),
+                     ycalib = data.frame(),
+                     paramCalib = data.frame(),
+                     paramsFormulas = list(),
+                     covFormulas = data.frame(),
                      surmodel = list(),
                      calib.opts = list()) {
   return(new("surcalib",
              yobs = yobs,
              xobs = xobs,
              xcalib = xcalib,
-             params.formulas = params.formulas,
-             cov.formulas = cov.formulas,
+             ycalib = ycalib,
+             paramCalib = paramCalib,
+             paramsFormulas = paramsFormulas,
+             covFormulas = covFormulas,
              surmodel = surmodel,
              calib.opts = calib.opts))
 }