+++ NEWS.md
@@ -1,6 +1,10 @@
 ## trollcalibr 0.0.9003
 
--   V1 
+-   V1: addExp, computeGP, computeHM, createExp, generateParam, lhscorcorr, processExp, 
+processSetup, processSim, setupDesign, setupExp, surcalib, surmodel, trolldae, trollexp,
+trollsetup, utilsStack, zzz
+
+Class added : trolldae, trollsetup, trollexpsingle, trollexpsetup
 
 
 ## trollcalibr 0.0.9002